+++ test/bract/ring/wrapper_test.clj
@@ -19,10 +19,13 @@
     [bract.ring.keydef      :as ring-kdef]
     [bract.ring.wrapper     :as wrapper])
   (:import
+    [java.io ByteArrayInputStream]
     [org.eclipse.jetty.server Server]))
 
 
-(def default-config (core-kdef/resolve-config {} ["bract/ring/config.edn"]))
+(def default-config (-> {}
+                      (core-kdef/resolve-config ["bract/ring/config.edn"])
+                      (assoc "bract.core.eventlog.enable" true)))
 
 
 (def default-response {:status 200
@@ -55,6 +58,21 @@
         (update-in [:headers] dissoc "Date" "Connection" "Server"))))
   ([handler jetty-options path]
     (roundtrip-get handler jetty-options path {})))
+
+
+(defn roundtrip-post
+  ([handler jetty-options path]
+    (roundtrip-post handler jetty-options path nil))
+  ([handler jetty-options path http-options]
+    (with-jetty handler (conj {:port 3000 :join? false}
+                          jetty-options)
+      (-> (str "http://localhost:3000" path)
+        (client/post (merge {:throw-exceptions false} http-options))
+        (dissoc
+          :trace-redirects :length :orig-content-encoding
+          :request-time :repeatable? :protocol-version
+          :streaming? :chunked? :reason-phrase)
+        (update-in [:headers] dissoc "Date" "Connection" "Server")))))
 
 
 (defn roundtrip-put
@@ -233,6 +251,18 @@
             (-> (roundtrip-get wrapped-handler {} "/ping")
               :body)
             (-> (roundtrip-get wrapped-handler {:async? true} "/ping")
+              :body)))
+      (is (= "song"
+            (-> (wrapped-handler {:uri "/ping" :request-method :post :body (-> "song"
+                                                                             (.getBytes)
+                                                                             (ByteArrayInputStream.))})
+              :body)
+            (-> (roundtrip-post wrapped-handler {} "/ping" {:body "song"
+                                                            ;; :headers {"Content-type" "text/plain"}
+                                                            })
+              :body)
+            (-> (roundtrip-post wrapped-handler {:async? true} "/ping" {:body "song"
+                                                                        :headers {"Content-type" "text/plain"}})
               :body))))
     (testing "ping URI with trailing slash"
       (is (not= default-response
@@ -243,6 +273,18 @@
             (-> (roundtrip-get wrapped-handler {} "/ping/")
               :body)
             (-> (roundtrip-get wrapped-handler {:async? true} "/ping/")
+              :body)))
+      (is (= "song"
+            (-> (wrapped-handler {:uri "/ping" :request-method :post :body (-> "song"
+                                                                             (.getBytes)
+                                                                             (ByteArrayInputStream.))})
+              :body)
+            (-> (roundtrip-post wrapped-handler {} "/ping" {:body "song"
+                                                            ;; :headers {"Content-type" "text/plain"}
+                                                            })
+              :body)
+            (-> (roundtrip-post wrapped-handler {:async? true} "/ping" {:body "song"
+                                                                        :headers {"Content-type" "text/plain"}})
               :body))))))
 
 
@@ -470,3 +512,62 @@ Every request must bear the header 'x-trace-id'"}
 Header 'x-trace-id' has invalid value: ID too short"}
           (roundtrip-get nlength {} "/" {:headers {"x-trace-id" "123"}})
           (roundtrip-get nlength {:async true} "/" {:headers {"x-trace-id" "123"}})))))
+
+
+(deftest test-traffic-log
+  (let [request-logger   (atom nil)
+        response-logger  (atom nil)
+        exception-logger (atom nil)
+        reset-loggers!   (fn []
+                           (reset! request-logger nil)
+                           (reset! response-logger nil)
+                           (reset! exception-logger nil))
+        happy-handler (fn self
+                        ([request]               {:status 200 :body "OK"})
+                        ([request respond raise] (respond (self request))))
+        error-handler (fn self
+                        ([request]               (throw (Exception. "test")))
+                        ([request respond raise] (raise (Exception. "test"))))
+        wrap-handler  (fn [f]
+                        (-> f
+                          (wrapper/traffic-log-wrapper
+                            {:bract.core/config
+                             {"bract.ring.traffic.log.enabled" true
+                              "bract.ring.traffic.log.options"
+                              {:request-logger (fn [request] (reset! request-logger :called))
+                               :response-logger (fn [request response ^double duration-millis]
+                                                  (reset! response-logger :called))
+                               :exception-logger (fn [request response ^double duration-millis]
+                                                   (reset! exception-logger :called))}}})))]
+    (testing "happy sync"
+      (reset-loggers!)
+      (roundtrip-get (wrap-handler happy-handler) {} "/" {})
+      (is (= :called
+            @request-logger
+            @response-logger))
+      (is (nil?
+            @exception-logger)))
+    (testing "error sync"
+      (reset-loggers!)
+      (roundtrip-get (wrap-handler error-handler) {} "/" {})
+      (is (= :called
+            @request-logger
+            @exception-logger))
+      (is (nil?
+            @response-logger)))
+    (testing "happy async"
+      (reset-loggers!)
+      (roundtrip-get (wrap-handler happy-handler) {:async true} "/" {})
+      (is (= :called
+            @request-logger
+            @response-logger))
+      (is (nil?
+            @exception-logger)))
+    (testing "error sync"
+      (reset-loggers!)
+      (roundtrip-get (wrap-handler error-handler) {:async true} "/" {})
+      (is (= :called
+            @request-logger
+            @exception-logger))
+      (is (nil?
+            @response-logger)))))